// src/modules/configuration/model/ConditionStatement.ts
export interface IConditionStatement {
    IDConditionStatement: number,
    ConditionStatement: string;
    FlowControlOperator: any;
    IDProcedureDestination: number;
    ConditionalRow: number;
    ConditionalValue: string;
    IDResponseValue: number;
    FormName: string;
    IDDocument: number;
    ConditionalOperator: string;
    IDProcedureSource: number;
}